+++ src/plugins/ui.h
@@ -8,6 +8,7 @@
 #include "ui/overlay.h"
 #include "ui/rendering.h"
 #include "ui/systems.h"
+#include "ui/text_measure.h"
 #include "ui/theme.h"
 #include "ui/theme_io.h"
 #include "ui/utilities.h"